test(epic): test the error catching setup
