+++ test/renderer/epics/config-spec.js
@@ -1,37 +1,13 @@
-import chai, { expect } from 'chai';
+import { expect } from 'chai';
 import { ActionsObservable } from 'redux-observable';
 
 import {
-  LOAD_CONFIG,
-  loadConfigEpic,
   saveConfigOnChangeEpic,
-  retryAndEmitError,
 } from '../../../src/notebook/epics/config';
-
-const sinon = require('sinon');
-const sinonChai = require('sinon-chai');
-
-chai.use(sinonChai);
 
 const Rx = require('rxjs/Rx');
 
 const Observable = Rx.Observable;
-
-describe('loadConfigEpic', () => {
-  it('errors on a bad read', (done) => {
-    const input$ = Observable.of({ type: LOAD_CONFIG }).share();
-    const action$ = new ActionsObservable(input$);
-    const responseActions = loadConfigEpic(action$);
-    responseActions.subscribe(
-      (x) => {
-        expect(x.type).to.equal('ERROR');
-        done();
-      },
-      expect.fail,
-      expect.fail,
-    );
-  });
-});
 
 describe('saveConfigOnChangeEpic', () => {
   it('invokes a SAVE when the SET_CONFIG_KEY action happens', (done) => {
@@ -46,25 +22,5 @@ describe('saveConfigOnChangeEpic', () => {
       expect.fail,
       expect.fail,
     );
-  });
-});
-
-describe('retryAndEmitError', () => {
-  it('returns the source observable, emitting an error action first', () => {
-    const source = {
-      startWith: sinon.stub(),
-    };
-    source.startWith.returns(source);
-    const err = new Error('Oh no!');
-    const newSource = retryAndEmitError(err, source);
-
-    expect(source.startWith.calledOnce).to.equal(true);
-    expect(source.startWith).to.have.been.calledWith({
-      payload: err,
-      error: true,
-      type: 'ERROR',
-    });
-
-    expect(newSource).to.equal(source);
   });
 });
